Optimize topksoftmax WARPS_PER_TB for higher occupancy and remove redundant precision conversion (#652)
* apply clang-format

* optimize

---------

+++ aiter/fused_moe.py
@@ -911,7 +911,7 @@ def fused_topk(
         topk_weights,
         topk_ids,
         token_expert_indicies,
-        gating_output.float(),  # TODO(woosuk): Optimize this.
+        gating_output,
         renormalize,
     )
     del token_expert_indicies  # Not used. Will be used in the future.
